+++ doctor_appointment_frontend/package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "react-router-dom": "^6.15.0",
     "@reduxjs/toolkit": "^1.9.5",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
@@ -15,6 +14,8 @@
     "react-router-dom": "^6.15.0",
     "react-scripts": "5.0.1",
     "react-transition-group": "^4.4.5",
+    "stylelint-csstree-validator": "^3.0.0",
+    "stylelint-scss": "^5.1.0",
     "svgo": "^3.0.2",
     "web-vitals": "^2.1.4"
   },
@@ -49,8 +50,6 @@
     "eslint-config-airbnb": "^19.0.4",
     "stylelint": "^15.10.3",
     "stylelint-config-standard": "^34.0.0",
-    "tailwindcss": "^3.3.3",
-    "stylelint-csstree-validator": "^1.9.0",
-    "stylelint-scss": "^3.21.0"
+    "tailwindcss": "^3.3.3"
   }
 }
