+++ doctor_appointment_frontend/src/components/detail.js
