--- doctor_appointment_frontend/src/components/myreservation.js
@@ -1,9 +1,0 @@
-function Myreservation() {
-  return (
-    <section>
-      Myreservation
-    </section>
-  );
-}
-
-export default Myreservation;
