--- doctor_appointment_frontend/src/components/reserveForm.js
@@ -1,9 +1,0 @@
-function ReserveForm() {
-  return (
-    <section>
-      Reserve Form
-    </section>
-  );
-}
-
-export default ReserveForm;
